Roll recent updates to GH for navigation fixes
See:
https://github.com/dart-lang/angular_analyzer_plugin/pull/697
https://github.com/dart-lang/angular_analyzer_plugin/pull/698

Closes #1781

PiperOrigin-RevId: 244388876
